+++ sensirion_i2c_sen66/device.py
@@ -7,7 +7,7 @@
 #
 # Generator:     sensirion-driver-generator 1.1.2
 # Product:       sen66
-# Model-Version: 1.5.0
+# Model-Version: 1.6.0
 #
 """
 The class Sen66DeviceBase implements the low level interface of the sensor.
@@ -20,15 +20,15 @@ from sensirion_driver_support_types.mixin_access import MixinAccess
 from sensirion_i2c_sen66.commands import (ActivateShtHeater, DeviceReset, DeviceStatus, GetAmbientPressure,
                                           GetCo2SensorAutomaticSelfCalibration, GetDataReady,
                                           GetNoxAlgorithmTuningParameters, GetProductName, GetSensorAltitude,
-                                          GetSerialNumber, GetVocAlgorithmState, GetVocAlgorithmTuningParameters,
-                                          PerformForcedCo2Recalibration, ReadAndClearDeviceStatus, ReadDeviceStatus,
-                                          ReadMeasuredRawValues, ReadMeasuredValuesAsIntegers,
-                                          ReadNumberConcentrationValuesAsIntegers, SetAmbientPressure,
-                                          SetCo2SensorAutomaticSelfCalibration, SetNoxAlgorithmTuningParameters,
-                                          SetSensorAltitude, SetTemperatureAccelerationParameters,
-                                          SetTemperatureOffsetParameters, SetVocAlgorithmState,
-                                          SetVocAlgorithmTuningParameters, StartContinuousMeasurement,
-                                          StartFanCleaning, StopMeasurement)
+                                          GetSerialNumber, GetShtHeaterMeasurements, GetVersion, GetVocAlgorithmState,
+                                          GetVocAlgorithmTuningParameters, PerformForcedCo2Recalibration,
+                                          ReadAndClearDeviceStatus, ReadDeviceStatus, ReadMeasuredRawValues,
+                                          ReadMeasuredValuesAsIntegers, ReadNumberConcentrationValuesAsIntegers,
+                                          SetAmbientPressure, SetCo2SensorAutomaticSelfCalibration,
+                                          SetNoxAlgorithmTuningParameters, SetSensorAltitude,
+                                          SetTemperatureAccelerationParameters, SetTemperatureOffsetParameters,
+                                          SetVocAlgorithmState, SetVocAlgorithmTuningParameters,
+                                          StartContinuousMeasurement, StartFanCleaning, StopMeasurement)
 
 from sensirion_i2c_sen66.result_types import (SignalCo2, SignalHumidity, SignalMassConcentrationPm10p0,
                                               SignalMassConcentrationPm1p0, SignalMassConcentrationPm2p5,
@@ -603,13 +603,40 @@ class Sen66DeviceBase:
         to reverse creep at high humidity.
         This command activates the SHT sensor heater with 200mW for 1s.
         The heater is then automatically deactivated again.
+        The "get_sht_heater_measurements" command can be used to check if the
+        heater has finished (firmware version >= 4.0).
         Wait at least 20s after this command before starting a measurement to get
         coherent temperature values (heating consequence to disappear).
 
         .. note::
             This command is only available in idle mode.
+            For firmware version < 4.0, wait for at least 1300ms before sending another
+            command, to ensure heating is finsihed.
         """
         transfer = ActivateShtHeater()
+        return execute_transfer(self._channel, transfer)
+
+    def get_sht_heater_measurements(self):
+        """
+        Get the measured values when the SHT sensor heating is triggerd. If the
+        heating is not finished, the returned humidity and temperature values
+        are 0x7FFF.
+
+        :return humidity:
+            Value is scaled with factor 100: RH [%] = value / 100
+            *Note: If this value is not available, 0x7FFF is returned.*
+        :return temperature:
+            Value is scaled with factor 200: T [°C] = value / 200
+            *Note: If this value is not available, 0x7FFF is returned.*
+
+        .. note::
+            This command is only availble in idle mode.
+            This command is only available for firmware version >= 4.0.
+            This command must be used after the "activate_sht_heater" command.
+            The get_sht_heater_measurements command can be queried every 0.05s to get
+            the measurements.
+        """
+        transfer = GetShtHeaterMeasurements()
         return execute_transfer(self._channel, transfer)
 
     def get_product_name(self):
@@ -633,6 +660,18 @@ class Sen66DeviceBase:
         """
         transfer = GetSerialNumber()
         return execute_transfer(self._channel, transfer)[0]
+
+    def get_version(self):
+        """
+        Gets the version information for the hardware, firmware and communication protocol.
+
+        :return firmware_major:
+            Firmware major version number.
+        :return firmware_minor:
+            Firmware minor version number.
+        """
+        transfer = GetVersion()
+        return execute_transfer(self._channel, transfer)
 
     def read_device_status(self):
         """